backend/pyproject.toml
[project]
name = "fmsi-un-recommendations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

requires-python = ">=3.12"
dependencies = [
    "openai>=1.108.1",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.9",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "fastembed>=0.3.1",
    "pydantic-ai>=0.7.0",
    "streamlit>=1.40.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "psycopg[binary]>=3.1.0",
    "loguru>=0.7.0",
    "sentence-transformers>=2.7.0",
]

[project.scripts]
app = "fmsi_un_recommendations:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
notebook = [
    "ipykernel>=6.30.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.prettier]
print-width = 120
prose-wrap = "preserve"
tab-width = 2
use-tabs = false
semi = true
single-quote = false
trailing-comma = "es5"
bracket-spacing = true
arrow-parens = "avoid"
end-of-line = "lf"

[tool.biome]
# Biome configuration for future UI development
# Handles JavaScript, TypeScript, JSX, TSX, and JSON
# Fast Rust-based alternative to ESLint + Prettier
linter-enabled = true
formatter-enabled = true
indent-style = "space"
indent-width = 2
line-width = 120
quote-style = "double"
semicolons = "always"
trailing-commas = "es5"
